% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_diann.R
\name{convert_diann_to_msstats_in}
\alias{convert_diann_to_msstats_in}
\title{Convert diann parquet to msstats_in format}
\usage{
convert_diann_to_msstats_in(
  diann_parquet,
  qvalue = 0.01,
  contaminant_prefix = "Cont"
)
}
\arguments{
\item{diann_parquet}{Path to diann parquet main report}

\item{qvalue}{FDR for precursors (default 0.01)}

\item{contaminant_prefix}{Prefix for contaminants (default is Cont)}
}
\value{
Data frame with reliable precursors and without contaminants and the following columns for msstats_input format
}
\description{
Converts the diann parquet to msstats_in format after filtering for contaminants and qvalue of precursors, then renames columns to prepare dataframe for functions like the extract_peptides_per_protein
}
