% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mzTab.R
\name{count_peptide_modifications}
\alias{count_peptide_modifications}
\title{Count peptide modifications and generate plots}
\usage{
count_peptide_modifications(
  mzt,
  contaminant_prefix = "CONTAMINANT_",
  plot_type = "mixed",
  output_jpg = NULL,
  jpg_width = 10,
  jpg_height = 8
)
}
\arguments{
\item{mzt}{Path to mzTab file}

\item{contaminant_prefix}{The prefix for contaminants, default is CONTAMINANT_}

\item{plot_type}{The type of plot to generate:
- `"separate"`: Plots the frequency of each modification individually.
- `"mixed"`: Plots the frequency of all combinations of modifications.}

\item{output_jpg}{Optional. Name of the output jpg image to save.}

\item{jpg_width}{Image width (default 10)}

\item{jpg_height}{Image heigth (default 8)}
}
\value{
The ggplot with PTM frequencies
}
\description{
Counts peptide modifications in an mzTab file and generates plots of their frequencies.
}
