% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_diann.R
\name{count_psms_diann}
\alias{count_psms_diann}
\title{Count PSMs per proteinGroup or gene in diann_dataframe}
\usage{
count_psms_diann(diann_dataframe, response_level = "protein")
}
\arguments{
\item{diann_dataframe}{The diann dataframe, prefiltered with filter_diann_parquet}

\item{response_level}{Specifies the level to count:
\itemize{
\item \code{"protein"} (default): Will count peptidoforms and PSMs per proteinGroup
\item \code{"gene"}: Will count peptidoforms and PSMs per gene
}}
}
\value{
Data frame with the following columns:
\itemize{
\item \verb{protein_Id or gene_Id}: Protein or gene accession, one per row
\item \code{PSMs}: Number of PSMs.
}
}
\description{
Extracts PSM count for each proteinGroup or gene in the diann_dataframe. PENDING to add peptidoform count, not urgent since I do not include variable PTMs for now
}
