% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prolfqua_functions.R
\name{create_lfqdata}
\alias{create_lfqdata}
\title{Create an LFQData object for prolfqua from msstats_in}
\usage{
create_lfqdata(
  data,
  contaminant_prefix = "CONTAMINANT_",
  response_level = "peptide",
  proteinId_column = "ProteinName",
  factor_column = "Condition",
  extra_factor = NULL,
  factor_depth = 1
)
}
\arguments{
\item{data}{Dataframe with peptide at msstats_in format or protein data}

\item{contaminant_prefix}{Prefix for contaminants, default is CONTAMINANT_}

\item{response_level}{Specifies the level to use:
\itemize{
\item \code{"peptide"} (default): Aggregates peptide-level intensities.
\item \code{"protein"}: Uses protein-level intensities.
}}

\item{proteinId_column}{Name of column with protein IDs (deafult is "ProteinName")}

\item{factor_column}{Name of column with the experimental/biological factor (default is "Condition")}

\item{extra_factor}{Optional. Name of additional factor for protein level data, batches for example.}

\item{factor_depth}{The level of experimental factors (default 1). Set to one in case you want to test the interaction with the extra factor.}
}
\value{
The LFQData object
}
\description{
Converts a data frame of msstats input format or a custom data frame to LFQData object for the prolfqua package. Sample names in must be in Reference column, response in Intensity column. Accepts up to 2 experimental factors.
}
