% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_msstats_in.R
\name{extract_peptides_per_protein}
\alias{extract_peptides_per_protein}
\title{Extract peptide sequences, count and indices per protein}
\usage{
extract_peptides_per_protein(
  msstats_input_path,
  fasta_path,
  contaminant_prefix = "CONTAMINANT",
  delimiter = ";",
  output_file = NULL
)
}
\arguments{
\item{msstats_input_path}{Path to msstats input style file from quantms (probably works for other programs as well?)}

\item{fasta_path}{Path to fasta file with protein sequences (can be gzipped)}

\item{contaminant_prefix}{Prefix for contaminants (default is CONTAMINANT)}

\item{delimiter}{Delimiter for peptides shared between different proteins (default is ;)}

\item{output_file}{Optional. Name of the output file (better to merge it with the output of parse_mzTab::extract_protein_stats first and save together)}
}
\value{
Data frame with peptide info (without PTMs) for each detected protein including:
}
\description{
Processes the msstats_in (formatted as input for msstats) file to extract peptide sequences, number and indexes of quantified peptides for each protein. Maybe too long and needs splitting
}
\details{
This function performs the following steps:
  - Reads the msstats input file and removes PTM information from peptide sequences
  - Filters the FASTA file to include only detected proteins from msstats_in
  - Extracts peptide sequences, counts, and positional indices for each protein.
}
