% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_diann.R
\name{filter_diann_parquet}
\alias{filter_diann_parquet}
\title{Filter diann parquet for 1\% precursor FDR and contaminants}
\usage{
filter_diann_parquet(diann_parquet, qvalue = 0.01, contaminant_prefix = "Cont")
}
\arguments{
\item{diann_parquet}{Path to diann parquet main report}

\item{qvalue}{FDR for precursors (default 0.01)}

\item{contaminant_prefix}{Prefix for contaminants (default is Cont)}
}
\value{
Data frame with reliable precursors and without contaminants
}
\description{
Filter diann parquet for 1\% precursor FDR and contaminants
}
