% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_protein_fasta.R
\name{filter_fasta_file_with_ids}
\alias{filter_fasta_file_with_ids}
\title{Filter FASTA with IDs}
\usage{
filter_fasta_file_with_ids(
  fasta_path,
  protein_ids,
  uniprot_fasta_header = FALSE
)
}
\arguments{
\item{fasta_path}{Path to the FASTA file}

\item{protein_ids}{Vector of IDs to keep}

\item{uniprot_fasta_header}{Logical. If TRUE, parse UniProt-style headers.}
}
\value{
Filtered AAStringSet
}
\description{
Filter FASTA with IDs
}
