% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prolfqua_functions.R
\name{filter_for_leading_protein}
\alias{filter_for_leading_protein}
\title{Filter protein groups to retain leading protein IDs}
\usage{
filter_for_leading_protein(
  mp_df,
  proteomicslfq_ids_to_keep,
  protein_separator = ";"
)
}
\arguments{
\item{mp_df}{A dataframe with proteinGroups in \code{protein_Id} column}

\item{proteomicslfq_ids_to_keep}{A vector containing the proteomicsLFQ leading proteins}

\item{delimiter}{Delimiter for multiple proteins in a protein group (default is ;)}
}
\value{
A dataframe with only leading protein in \code{protein_Id} column
}
\description{
For protein groups with multiple IDs, filters the data frame to retain only the leading protein ID based on a given list.
}
