% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mzTab.R
\name{get_proteinIds_from_proteomicslfq}
\alias{get_proteinIds_from_proteomicslfq}
\title{Extract leading protein IDs from mzTab}
\usage{
get_proteinIds_from_proteomicslfq(mzt, contaminant_prefix = "CONTAMINANT_")
}
\arguments{
\item{mzt}{Path to mzTab file}

\item{contaminant_prefix}{The prefix for contaminants, default is CONTAMINANT_}
}
\value{
Vector with one protein per proteinGroug
}
\description{
Extracts a vector of protein IDs from the mzTab file. This file has one id per protein group.
}
