% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{get_valid_ids}
\alias{get_valid_ids}
\title{Get Valid Protein IDs}
\usage{
get_valid_ids(all_ids, valid_ids, delimiter = ";")
}
\arguments{
\item{all_ids}{A vector with one or multiple IDs per item}

\item{valid_ids}{A vector with the IDs to keep}

\item{delimiter}{The delimiter in case of multiple IDs per row (default is ;)}
}
\value{
The filtered vector with one leading ID per group
}
\description{
Filter vector of IDs, retaining only main IDs  (as determined from proteomicsLFQ mzTab) in case of proteinGroups
}
