% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_protein_fasta.R
\name{make_decoy_AAStringSet}
\alias{make_decoy_AAStringSet}
\title{Generate pseudo-reversed decoy sequences as Biostrings AAStringSet}
\usage{
make_decoy_AAStringSet(fasta_path, uniprot_fasta_header = FALSE)

make_decoy_AAStringSet(fasta_path, uniprot_fasta_header = FALSE)
}
\arguments{
\item{fasta_path}{Path to the FASTA file}

\item{uniprot_fasta_header}{Logical. If TRUE, parse UniProt-style headers.}
}
\value{
AAStringSet of decoy sequences prefixed by "DECOY_"

AAStringSet of reversed decoy protein sequences, with headers prefixed by "DECOY_"
}
\description{
Keeps the first residue fixed (methionine) and reverses remaining sequence.
}
