% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_protein_fasta.R
\name{read_and_parse_fasta}
\alias{read_and_parse_fasta}
\title{Read protein FASTA and parse headers}
\usage{
read_and_parse_fasta(fasta_path, uniprot_fasta_header = FALSE)
}
\arguments{
\item{fasta_path}{Path to the FASTA file}

\item{uniprot_fasta_header}{Logical. If TRUE, parse UniProt-style headers.}
}
\value{
Biostrings AAStringSet with parsed names
}
\description{
Reads a FASTA file as an AAStringSet and simplifies the sequence names:
\itemize{
\item Removes everything after the first space.
\item If UniProt-style headers, extracts the accession.
}
}
