% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_diann.R
\name{rename_diann_cols_like_msstats_in}
\alias{rename_diann_cols_like_msstats_in}
\title{Rename columns of diann dataframe to match msstats_in names}
\usage{
rename_diann_cols_like_msstats_in(diann_dataframe)
}
\arguments{
\item{diann_dataframe}{The diann dataframe, prefiltered with filter_diann_parquet}
}
\value{
Data frame with reliable precursors and without contaminants and the following columns for msstats_input format
}
\description{
Converts the diann parquet to msstats_in format by renaming columns to prepare dataframe for functions like the extract_peptides_per_protein
}
