% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{save_gzipped_csv}
\alias{save_gzipped_csv}
\title{Save gzipped csv file}
\usage{
save_gzipped_csv(df, output_file, row_names = FALSE)
}
\arguments{
\item{df}{The dataframe}

\item{output_file}{The name of the output fle, if given in a directory that does not exist it will be created}

\item{row_names}{Logical; whether to include row names in the output (default is FALSE)}
}
\value{
None
}
\description{
Saves a dataframe to a gzipped csv file, optionally including row names. . If the specified directory does not exist, it will be created.
}
