% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_msstats_in.R
\name{extract_peptides_per_protein}
\alias{extract_peptides_per_protein}
\title{Extract peptide sequences, count and indices per protein}
\usage{
extract_peptides_per_protein(msstats_in_df, fasta_path, delimiter = ";")
}
\arguments{
\item{msstats_in_df}{msstats_input style dataframe with PeptideSequence column stripped from PTMs and proteinId in column ProteinName}

\item{fasta_path}{Path to fasta file with protein sequences (can be gzipped)}

\item{delimiter}{Delimiter for peptides shared between different proteins (default is ;)}
}
\value{
Data frame with peptide info (without PTMs) for each detected protein including:
}
\description{
Processes the msstats_in_df dataframe to extract peptide sequences, number and indexes of quantified peptides for each protein
}
\details{
This function performs the following steps:
\itemize{
\item Filters the FASTA file to include only detected proteins from msstats_in_df
\item Extracts peptide sequences, counts, and positional indices for each protein.
}
}
