% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prolfqua_functions.R
\name{filter_proteins_by_replicates}
\alias{filter_proteins_by_replicates}
\title{Filter proteins by replicate counts per condition}
\usage{
filter_proteins_by_replicates(mp_df, conditions, min_replicates)
}
\arguments{
\item{mp_df}{The dataframe with protein quantifications}

\item{conditions}{The vector of the conditions to search the mp_df colnames}

\item{min_replicates}{Minimum number of min replicates needed}
}
\value{
The dataframe with proteins passing the filtering
}
\description{
Retains proteins that are quantified in at least a specified number of replicates for one condition.
}
