% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prolfqua_functions.R
\name{protein_aggregation}
\alias{protein_aggregation}
\title{Aggregate peptide intensities to protein level}
\usage{
protein_aggregation(lfqdataPeptideNorm)
}
\arguments{
\item{lfqdataPeptideNorm}{The prolfqua object after peptide normalization}
}
\value{
The a dataframe with protein intensities
}
\description{
Function for protein aggregation of peptide normalized data
\itemize{
\item Use Tukey's median polish protein aggregation
\item Remove artifacts with same value in all samples
\item Replace the artifact intensities with top3 method
\item Merge the two methods
}
}
